requirements.txt
streamlit==1.29.0
pandas==2.1.4
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
sentence-transformers==2.5.1
Pillow==10.1.0
scikit-learn==1.3.2
numpy==1.26.2

